+++ relayer/middleware/index.ts
@@ -6,4 +6,5 @@ export * from "./source-tx.middleware";
 export * from "./staging-area.middleware";
 export * from "./tokenBridge.middleware";
 export * from "./wallet";
+export * from "./missedVaasV3";
 export * as LegacyPluginCompat from "./legacy-plugin";
