+++ relayer/middleware/missedVaas/index.ts
@@ -1,3 +1,2 @@
 export { missedVaas } from "./missedVaas.middleware";
-export { spawnMissedVaaWorker } from "./missedVaasV3";
 export type { MissedVaaOpts } from "./missedVaas.middleware";
