+++ relayer/middleware/missedVaasV3/index.ts
@@ -1,0 +1,2 @@
+export { spawnMissedVaaWorker } from "./worker";
+export type { MissedVaaOpts as MissedVaaWorkerOpts } from "./worker";
